# api/server/sdk/api/api.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "apiSdkCloudBackupCatalogResponse": {
      "properties": {
        "contents": {
          "items": {
            "type": "string"
          },
          "title": "Contents is listing of backup contents",
          "type": "array"
        }
      },
      "title": "Defines a response containing the contents of a backup stored by a cloud provider",
      "type": "object"
    },
...
    "apiSdkCloudBackupHistoryItem": {
      "properties": {
        "src_volume_id": {
          "title": "SrcVolumeID is volume ID which was backedup",
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/apiSdkCloudBackupStatusType",
          "title": "Status indicates whether backup was completed/failed"
        },
        "timestamp": {
          "format": "date-time",
          "title": "TimeStamp is the time at which either backup completed/failed",
          "type": "string"
        }
      },
      "title": "SdkCloudBackupHistoryItem contains information about a backup for a\nspecific volume",
      "type": "object"
    },
    "apiSdkCloudBackupHistoryResponse": {
      "properties": {
        "history_list": {
          "items": {
            "$ref": "#/definitions/apiSdkCloudBackupHistoryItem"
          },
          "title": "HistoryList is list of past backups on this volume",
          "type": "array"
        }
      },
      "title": "Defines a response containing a list of history of backups to a cloud provider",
      "type": "object"
    },
...
    "apiSdkCloudBackupStatusType": {
      "default": "SdkCloudBackupStatusTypeUnknown",
      "description": "- SdkCloudBackupStatusTypeUnknown: Unkonwn\n - SdkCloudBackupStatusTypeNotStarted: Not started\n - SdkCloudBackupStatusTypeDone: Done\n - SdkCloudBackupStatusTypeAborted: Aborted\n - SdkCloudBackupStatusTypePaused: Paused\n - SdkCloudBackupStatusTypeStopped: Stopped\n - SdkCloudBackupStatusTypeActive: Active\n - SdkCloudBackupStatusTypeFailed: Failed",
      "enum": [
        "SdkCloudBackupStatusTypeUnknown",
        "SdkCloudBackupStatusTypeNotStarted",
        "SdkCloudBackupStatusTypeDone",
        "SdkCloudBackupStatusTypeAborted",
        "SdkCloudBackupStatusTypePaused",
        "SdkCloudBackupStatusTypeStopped",
        "SdkCloudBackupStatusTypeActive",
        "SdkCloudBackupStatusTypeFailed"
      ],
      "title": "CloudBackup status types",
      "type": "string"
    },
...
  },
  "info": {
    "title": "OpenStorage SDK",
    "version": "0.14.0"
  },
  "paths": {
...
    "/v1/cloudbackups/backup/catalog": {
      "get": {
        "operationId": "Catalog",
        "parameters": [
          {
            "description": "Id of the backup.",
            "in": "query",
            "name": "backup_id",
            "required": false,
            "type": "string"
          },
          {
            "description": "Credential id describe the credentials for the cloud.",
            "in": "query",
            "name": "credential_id",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiSdkCloudBackupCatalogResponse"
            }
          }
        },
        "summary": "Catalog returns a list of the contents in the backup",
        "tags": [
          "OpenStorageCloudBackup"
        ]
      }
    },
...
    "/v1/cloudbackups/volume/history": {
      "get": {
        "operationId": "History",
        "parameters": [
          {
            "description": "This optional value defines which history of backups is being\nrequested. If not provided, it will return the history for all volumes.",
            "in": "query",
            "name": "src_volume_id",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiSdkCloudBackupHistoryResponse"
            }
          }
        },
        "summary": "History returns a list of backups for a specified volume",
        "tags": [
          "OpenStorageCloudBackup"
        ]
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0"
}